Backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
blis==1.3.3
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
cryptography==46.0.5
cymem==2.0.13
dataclasses-json==0.6.7
distro==1.9.0
Django==5.2.12
django-cors-headers==4.9.0
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.2.0
gunicorn==25.1.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.23.5
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain-classic==1.0.2
langchain-community==0.4.1
langchain-core==1.2.18
langchain-experimental==0.4.1
langchain-huggingface==1.2.1
langchain-ollama==1.0.1
langchain-openai==1.1.11
langchain-pinecone==0.2.13
langchain-text-splitters==1.1.1
langsmith==0.7.17
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
murmurhash==1.0.15
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.3
ollama==0.6.1
openai==2.26.0
orjson==3.11.7
packaging==24.2
phonenumbers==9.0.25
pillow==12.1.1
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
preshed==3.0.12
presidio_analyzer==2.2.361
presidio_anonymizer==2.2.361
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pypdf==6.8.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==3.0.1
shellingham==1.5.4
simsimd==6.5.16
six==1.17.0
smart_open==7.5.1
sniffio==1.3.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.48
sqlparse==0.5.5
srsly==2.5.2
sympy==1.14.0
tenacity==9.1.4
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.12.0
tldextract==5.3.1
tokenizers==0.19.1
torch==2.10.0
tqdm==4.67.3
transformers==4.41.2
typer==0.24.1
typer-slim==0.24.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
wasabi==1.1.3
weasel==0.4.3
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
